+++ app/src/main/java/c0d3/vitreen/app/models/dto/sdto/ProductSDTO.kt
@@ -5,5 +5,5 @@ data class ProductSDTO(
     val title: String,
     val category: String,
     val location: String,
-    val price: Long
+    val price: Double
 )
